requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
analytics-python=1.4.0=pypi_0
anyio=3.5.0=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asgiref=3.5.0=pypi_0
astor=0.8.1=pypi_0
asttokens=2.0.5=pypi_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.2=pypi_0
asynctest=0.13.0=pypi_0
attrs=21.4.0=pypi_0
backoff=1.10.0=pypi_0
bcrypt=3.2.0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
bottleneck=1.3.2=py37h2a96729_1
ca-certificates=2022.2.1=haa95532_0
cached-property=1.5.2=pypi_0
certifi=2021.10.8=py37haa95532_2
cffi=1.15.0=pypi_0
charset-normalizer=2.0.12=pypi_0
click=8.0.4=pypi_0
colorama=0.4.4=pypi_0
cryptography=36.0.1=pypi_0
cycler=0.11.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
django=3.2.13=pypi_0
docopt=0.6.2=pypi_0
entrypoints=0.4=pypi_0
fastapi=0.75.0=pypi_0
ffmpy=0.3.0=pypi_0
frozenlist=1.3.0=pypi_0
gast=0.2.2=pypi_0
google-pasta=0.2.0=pypi_0
gradio=2.8.7=pypi_0
grpcio=1.44.0=pypi_0
h11=0.13.0=pypi_0
h5py=2.10.0=pypi_0
icc_rt=2019.0.0=h0cc432a_1
idna=3.3=pypi_0
imageio=2.16.0=pypi_0
importlib-metadata=4.11.1=pypi_0
importlib-resources=5.4.0=pypi_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=5.5.6=pypi_0
ipython=6.2.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.5=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.0.3=pypi_0
joblib=1.1.0=pyhd3eb1b0_0
jsonschema=4.4.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.1.2=pypi_0
jupyter-console=5.2.0=pypi_0
jupyter-core=4.9.2=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
keras=2.3.1=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.2=pypi_0
linkify-it-py=1.0.3=pypi_0
markdown=3.3.6=pypi_0
markdown-it-py=2.0.1=pypi_0
markupsafe=2.1.0=pypi_0
matplotlib=3.3.4=pypi_0
mdit-py-plugins=0.3.0=pypi_0
mdurl=0.1.0=pypi_0
mindspore=1.6.1=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py37h2bbff1b_0
mkl_fft=1.3.1=py37h277e83a_0
mkl_random=1.2.2=py37hf11a4ad_0
monotonic=1.6=pypi_0
multidict=6.0.2=pypi_0
mysqlclient=2.1.0=pypi_0
nbclient=0.5.11=pypi_0
nbconvert=6.4.2=pypi_0
nbformat=5.1.3=pypi_0
nest-asyncio=1.5.4=pypi_0
networkx=2.6.3=pypi_0
notebook=6.4.8=pypi_0
numexpr=2.8.1=py37hb80d3ca_0
numpy=1.20.1=pypi_0
numpy-base=1.21.5=py37hc2deb75_0
opencv-python=4.5.1.48=pypi_0
openssl=1.1.1m=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
orjson=3.6.7=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.4=py37h6214cd6_0
pandocfilters=1.5.0=pypi_0
paramiko=2.9.2=pypi_0
parso=0.8.3=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=pypi_0
pip=21.2.4=py37haa95532_0
prometheus-client=0.13.1=pypi_0
prompt-toolkit=1.0.18=pypi_0
protobuf=3.19.4=pypi_0
psutil=5.9.0=pypi_0
pycparser=2.21=pypi_0
pycryptodome=3.14.1=pypi_0
pydantic=1.9.0=pypi_0
pydicom=2.1.2=pypi_0
pydub=0.25.1=pypi_0
pygments=2.11.2=pypi_0
pykwalify=1.8.0=pypi_0
pynacl=1.5.0=pypi_0
pynrrd=0.4.2=pypi_0
pyparsing=3.0.7=pypi_0
pyradiomics=3.0.1.post13+g2e0b76e=pypi_0
pyrsistent=0.18.1=pypi_0
python=3.7.5=h8c8aaf0_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-multipart=0.0.5=pypi_0
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.2.0=pypi_0
pywin32=303=pypi_0
pywinpty=2.0.4=pypi_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=pypi_0
qtconsole=5.2.2=pypi_0
qtpy=2.0.1=pypi_0
requests=2.27.1=pypi_0
ruamel-yaml=0.17.21=pypi_0
ruamel-yaml-clib=0.2.6=pypi_0
scikit-image=0.18.1=pypi_0
scikit-learn=1.0.2=py37hf11a4ad_1
scipy=1.6.2=pypi_0
send2trash=1.8.0=pypi_0
setuptools=58.0.4=py37haa95532_0
simplegeneric=0.8.1=pypi_0
simpleitk=2.1.1=pypi_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=pypi_0
sqlite=3.37.2=h2bbff1b_0
sqlparse=0.4.2=pypi_0
starlette=0.17.1=pypi_0
tensorboard=1.15.0=pypi_0
tensorflow=1.15.0=pypi_0
tensorflow-estimator=1.15.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.13.2=pypi_0
testpath=0.6.0=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2021.11.2=pypi_0
tornado=6.1=pypi_0
traitlets=5.1.1=pypi_0
typing-extensions=4.1.1=pypi_0
uc-micro-py=1.0.1=pypi_0
urllib3=1.26.8=pypi_0
uvicorn=0.17.5=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=pypi_0
wincertstore=0.2=py37haa95532_2
wrapt=1.13.3=pypi_0
yarl=1.7.2=pypi_0
zipp=3.7.0=pypi_0
